[Fix] ServerAction separation
