+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,5 @@
 import NextAuth from 'next-auth';
-import { authOptions } from '../\bauthOptions';
+import { authOptions } from '../authOptions';
 
 const handler = NextAuth(authOptions);
