+++ src/app/service/auth.ts
@@ -1,7 +1,5 @@
 import API from '@/utils/axios';
 import { FormValues } from '@/app/(auth)/components/signup/SignupForm';
-import { cookies } from 'next/headers';
-import { RequestCookie } from 'next/dist/compiled/@edge-runtime/cookies';
 
 interface ICredentials {
   email: string;
@@ -51,24 +49,4 @@ export const validateNickname = async (username: string) => {
 export const signUp = async (formValues: FormValues) => {
   const res = await API.post('/member/register', formValues);
   return res.data;
-};
-
-// AccessToken 재발급
-export const getNewAccessToken = async (accessToken: string) => {
-  try {
-    const { value } = cookies().get('refresh-token') as RequestCookie;
-
-    return await API.get('/member/token', {
-      headers: {
-        cookie: `access-token=${accessToken}; refresh-token=${value}`,
-      },
-    });
-  } catch (e) {
-    console.error(e);
-  }
-};
-
-//쿠키 삭제(server-action)
-export const deleteCookie = (name: string) => {
-  cookies().delete(name);
 };
